% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc692s.R
\name{jc69.2s}
\alias{jc69.2s}
\alias{jc69.2smle}
\alias{jc69.2slnL}
\alias{jc69.2slnP}
\title{The JC69 model on a pairwise sequence alignment}
\usage{
jc69.2smle(x, n)

jc69.2slnL(d, x, n)

jc69.2slnP(d, x, n, shape, rate, C = 1)
}
\arguments{
\item{x}{numeric, the number of differences in the alignment}

\item{n}{numeric, the number of sites in the alignment}

\item{d}{numeric, the molecular distance}

\item{shape}{numeric, the shape of the gamma prior on d}

\item{rate}{numeric, the rate of the gamma prior on d}

\item{C}{numeric, a scaling constant}
}
\value{
\code{jc69.2slnL} and \code{jc69.2slnP} return vectors of
  log-likelihood and log-posterior values respectively.
}
\description{
The JC69 model on a pairwise sequence alignment
}
\details{
\code{jc692s.mle} calulates the MLE of d. \code{jc69.2slnL} computes
  the log-likelihood and \code{jc69.2slnP} computes the log-posterior
  assuming a gamma prior on \code{d}.
}
\examples{
data(humanorang)

par(mfrow=c(1,2))
# Plot the likelihood:
curve(exp(jc69.2slnL(x, x=humanorang$x, n=humanorang$n)),
  n=5e2, from=0, to=0.2, ylab="Likelihood", xlab="Molecular distance (d)")
abline(v=jc69.2smle(x=humanorang$x, n=humanorang$n), col="red", lty=2)

# Plot the posterior:
curve(exp(jc69.2slnP(x, x=humanorang$x, n=humanorang$n, shape=1, rate=5)),
  n=5e2, from=0, to=0.2, ylab="Posterior", xlab="Molecular distance (d)")

# Does not integrate to one:
integrate(function(x) exp(jc69.2slnP(x, x=humanorang$x, n=humanorang$n,
  shape=1, rate=5)), lower=0, upper=Inf, abs.tol=0)

# Integrates to one:
integrate(function(x) exp(jc69.2slnP(x, x=humanorang$x, n=humanorang$n,
  shape=1, rate=5, C=1/5.167762e-131)), lower=0, upper=Inf, abs.tol=0)

}
\references{
Yang Z (2014) \emph{Molecular evolution: A Statistical approach},
  Oxford University Press.
}
\author{
Mario dos Reis
}
